refactor(options): :truck: move classes to own files
